Refactor: Consolidate atomic operations and worker optimization tests
